linting - remove extra exports from component files
