+++ src/store/appcontext.tsx
@@ -1,6 +1,7 @@
 import { ReactNode, createContext, useState } from 'react';
-import { DeckState, MAX_REELS, TileKeyCollection } from './data';
+import { DeckState, MAX_REELS, INITIAL_TOKENS, MAX_REEL_TOKENS, UiState, TileKeyCollection } from './data';
 import { clamp } from '../utils';
+import { ReelState, insertAfterPosition, insertIntoArray, removeAtPosition } from './utils';
 
 const AppContext = createContext({} as AppContextType);
 interface AppContextType {
@@ -30,48 +31,6 @@ interface AppContextType {
   insertReel: (positionIdx: number) => void;
 }
 
-export type ReelState = string[];
-
-export const insertIntoArray = (positionIdx: number, newItem: any, array: any[]) => {
-  const newPos = positionIdx + 1;
-
-  if (newPos < 0 || newPos > array.length) {
-    console.error(`invalid index ${positionIdx} provided`);
-    return array;
-  }
-
-  return [...array.slice(0, newPos), newItem, ...array.slice(newPos)];
-};
-
-export const insertAfterPosition = (reelIdx: number, positionIdx: number, tileKey: string, reelStates: ReelState[]) => {
-  return reelStates.map((reelState, rIdx) => {
-    if (rIdx === reelIdx) {
-      return insertIntoArray(positionIdx, tileKey, reelState);
-    } else {
-      return reelState;
-    }
-  });
-};
-
-export const removeAtPosition = (reelIdx: number, positionIdx: number, reelStates: ReelState[]) => {
-  if(reelStates.length === 1 && reelStates[0].length === 1){
-    console.log('ARE YOU CRAZY?!?! YOU CANT HAVE NOTHING!!!!!');
-    return reelStates;
-  }
-
-  return reelStates.map((reelState, rIdx) => {
-    if(rIdx === reelIdx){
-      return reelState.filter((_, index) => index !== positionIdx);
-    } else {
-      return reelState;
-    }
-  }).filter(rs => rs.length > 0);
-}
-
-export const INITIAL_TOKENS = 2;
-export const MAX_REEL_TOKENS = 5;
-
-export type UiState = 'game' | 'editor';
 interface Props {
   children: ReactNode;
 }
