+++ src/store/data.tsx
@@ -5,6 +5,11 @@ export const REEL_HEIGHT = 120; // height of each reel cell
 export const REEL_OVERLAP = 2; // # of looparound cells to add to edge of reel so that it can transition nicely
 export const SPIN_POWER_RANGE: MinMaxTouple = [0.01, 0.03]; // RNG speed range for each reel
 export const MAX_REELS = 6;
+export const INITIAL_TOKENS = 2;
+export const MAX_REEL_TOKENS = 5;
+export const MAX_HAND_SIZE = 3;
+
+export type UiState = 'game' | 'editor';
 
 /**
  * Tiles are defined in the TileGlossary (unique)
